cr30reader/pyproject.toml
[tool.poetry]
name = "cr30reader"
version = "0.1.0"
description = "Color Chart Reader for CR30 Colorimeter"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "cr30reader", from = "."}]

[tool.poetry.dependencies]
python = "^3.8"
pyserial-asyncio = ">=0.6"
numpy = ">=1.21.0"
matplotlib = ">=3.5.0"
colour-science = ">=0.4.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"
pytest-cov = "^4.0.0"

[tool.poetry.scripts]
run = "cr30reader.gui:main"
test = "cr30reader.tests.__main__:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

